+++ defaults/controllers/_default/root/admin/permissions/_index.js
@@ -1,0 +1,12 @@
+// --------------------
+// Permissions resource controller
+// Route definition
+// --------------------
+
+// exports
+
+// route definition
+exports = module.exports = {
+	// route params
+	type: 'resource'
+};