+++ defaults/controllers/_default/root/admin/roles/_index.js
@@ -8,6 +8,5 @@
 // route definition
 exports = module.exports = {
 	// route params
-	type: 'resource',
-	parentType: 'namespace'
+	type: 'resource'
 };